entry label
I add the textvariable part of the entry label for both user and processed emotion.  Just need to add the variable user and processed emotion label.  Not sure if we will use the first one but it's there.  I still need to add this variable emotionPrediction to the function that saves the processes emotion from the NN after it saves the prediction. Was unclear where that is in the program?